Quiz/requirements.txt
absl-py==2.0.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
breadability==0.1.20
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cors==1.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==4.4.2
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.0.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.5.4
gevent==24.2.1
google-ai-generativelanguage==0.4.0
google-api-core==2.19.2
google-auth==2.34.0
google-generativeai==0.3.1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.0.3
groq==0.4.2
grpcio==1.66.1
grpcio-status==1.62.3
gTTS==2.5.3
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.23.4
idna==3.4
imageio==2.35.1
imageio-ffmpeg==0.5.1
intel-openmp==2021.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.6.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_core==5.7.1
keras==3.4.1
kiwisolver==1.4.5
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.18
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
langsmith==0.1.85
libclang==16.0.6
llama-index==0.10.18
llama-index-agent-openai==0.1.7
llama-index-cli==0.1.13
llama-index-core==0.10.68.post1
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-openai==0.1.11
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48.post3
llama-index-llms-groq==0.1.3
llama-index-llms-openai==0.1.31
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.9
llama-index-program-openai==0.1.7
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.33
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.9
llamaindex-py-client==0.1.19
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
minijinja==2.2.0
mkl==2021.4.0
ml-dtypes==0.3.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==1.52.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.4
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.6
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.24.4
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==24.6.1
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.1.2
pypdf==4.3.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-file==2.1.0
rich==13.7.1
rpds-py==0.17.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
selenium==4.16.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
striprtf==0.0.26
sumy==0.11.0
sympy==1.13.1
tbb==2021.13.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.1
tensorflow==2.16.2
tensorflow-estimator==2.14.0
tensorflow-intel==2.16.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tf_keras==2.16.0
threadpoolctl==3.2.0
tiktoken==0.8.0
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.5.0+cpu
torchaudio==2.5.0+cpu
torchvision==0.20.0+cpu
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.42.4
trio==0.23.2
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
ultralytics==8.2.75
ultralytics-thop==2.0.0
urllib3==2.0.6
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.0
wrapt==1.14.1
wsproto==1.2.0
xlwings==0.31.6
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==7.0.3
